requirements.txt
chrisapp~=2.5.3
nose~=1.3.7
pydicom
SimpleITK
numpy
pillow
